🎉 Add: wishService 생성
